requirements.txt

requests
google-cloud-vision
